move event status to a new file
